+++ rdsa_utils/helpers/python.py
@@ -4,9 +4,12 @@ import itertools
 import json
 import logging
 from datetime import datetime, time
+from functools import reduce, wraps
+from itertools import tee
 from typing import Any, Callable, Dict, Iterable, List, Mapping, Tuple, Union
 
 import pandas as pd
+from codetiming import Timer
 from more_itertools import always_iterable
 from pandas.tseries.offsets import MonthEnd
 
@@ -303,3 +306,351 @@ def convert_date_strings_to_datetimes(
     # Ensure dates are timestamp to enable inclusive filtering of provided end
     # date, see https://stackoverflow.com/a/43403904 for info.
     return (pd.Timestamp(start_date), pd.Timestamp(end_date))
+
+
+def time_it(*timer_args, **timer_kwargs) -> Callable:
+    """Measure the execution time of a function, with options to configure Timer.
+
+    Parameters
+    ----------
+    timer_args
+        Positional arguments to pass to the Timer object.
+    timer_kwargs
+        Keyword arguments to pass to the Timer object.
+
+    Returns
+    -------
+    Callable
+        A wrapped function that includes timing measurement.
+
+    Example
+    -------
+    @time_it()
+    def example_function():
+        # Function implementation
+    """
+
+    def decorator(func: Callable) -> Callable:
+        @wraps(func)
+        def wrap(*args, **kwargs) -> Any:
+            with Timer(*timer_args, **timer_kwargs) as t:
+                result = func(*args, **kwargs)
+            logger.info(f"<Executed {func.__name__} in {t.last:.2f} seconds>")
+            return result
+
+        return wrap
+
+    return decorator
+
+
+def setdiff(a: Iterable, b: Iterable) -> List[Any]:
+    """Return a list of elements that are present in `a` but not in `b`.
+
+    Parameters
+    ----------
+    a
+        The first iterable from which elements are to be selected.
+    b
+        The second iterable containing elements to be excluded.
+
+    Returns
+    -------
+    list
+        A list of elements that are in `a` but not in `b`.
+
+    Examples
+    --------
+    >>> setdiff([1, 2, 3, 4], [3, 4, 5, 6])
+    [1, 2]
+    >>> setdiff('abcdef', 'bdf')
+    ['a', 'c', 'e']
+    >>> setdiff({1, 2, 3}, {2, 3, 4})
+    [1]
+    >>> setdiff(range(5), range(2, 7))
+    [0, 1]
+    """
+    if not isinstance(a, Iterable) or not isinstance(b, Iterable):
+        msg = "Both inputs must be iterable."
+        raise TypeError(msg)
+
+    return list(set(a) - set(b))
+
+
+def flatten_iterable(
+    iterable: Iterable,
+    types_to_flatten: Union[type, Tuple] = (list, tuple),
+) -> List:
+    """Flatten an iterable.
+
+    Parameters
+    ----------
+    iterable
+        An iterable that may contain elements of various types.
+    types_to_flatten
+        Data type(s) that should be flattened. Defaults to (list, tuple).
+
+    Returns
+    -------
+    list
+        A flattened list with all elements from the input iterable, with
+        specified types unpacked.
+
+    Examples
+    --------
+    >>> flatten_iterable([1, [2, 3], (4, 5), 'abc'])
+    [1, 2, 3, 4, 5, 'abc']
+    >>> flatten_iterable([1, [2, 3], (4, 5), 'abc'], types_to_flatten=list)
+    [1, 2, 3, (4, 5), 'abc']
+    >>> flatten_iterable(['a', 'bc', ['d', 'e']], types_to_flatten=str)
+    ['a', 'b', 'c', 'd', 'e']
+    >>> flatten_iterable((1, [2, 3], (4, 5), 'abc'), types_to_flatten=(list, tuple))
+    (1, 2, 3, 4, 5, 'abc')
+    """
+    if not hasattr(iterable, "__iter__"):
+        msg = "`iterable` must be an iterable."
+        raise TypeError(msg)
+
+    if not isinstance(types_to_flatten, (type, tuple)):
+        msg = "`types_to_flatten` must be a type or a tuple of types."
+        raise TypeError(msg)
+
+    if isinstance(types_to_flatten, tuple):
+        if not all(isinstance(t, type) for t in types_to_flatten):
+            msg = "All elements in `types_to_flatten` must be types."
+            raise ValueError(msg)
+    else:
+        if not isinstance(types_to_flatten, type):
+            msg = "`types_to_flatten` must be a type or a tuple of types."
+            raise TypeError(msg)
+
+    flattened = []
+    for item in iterable:
+        if isinstance(item, types_to_flatten):
+            flattened.extend(item)
+        else:
+            flattened.append(item)
+    return flattened
+
+
+def convert_types_iterable(lst: Iterable, dtype: type = float) -> List:
+    """Convert the data type of elements in an iterable.
+
+    Parameters
+    ----------
+    lst
+        The iterable whose elements are to be converted.
+    dtype
+        The target data type to which elements in the iterable should be
+        converted. Defaults to `float`.
+
+    Returns
+    -------
+    list
+        A new list with elements converted to the specified data type.
+
+    Examples
+    --------
+    >>> convert_types_iterable([1, 2, 3])
+    [1.0, 2.0, 3.0]
+
+    >>> convert_types_iterable((10, 20, 30), dtype=str)
+    ['10', '20', '30']
+
+    >>> convert_types_iterable({'a', 'b', 'c'}, dtype=ord)
+    [97, 98, 99]
+
+    >>> convert_types_iterable(['10', '20', '30'], dtype=int)
+    [10, 20, 30]
+    """
+    if not isinstance(lst, (list, tuple, set, frozenset, range)):
+        msg = (
+            "Input must be an iterable type such as list, tuple, set, "
+            "frozenset, or range."
+        )
+        raise TypeError(msg)
+
+    if not isinstance(dtype, type):
+        msg = "`dtype` must be a valid type."
+        raise TypeError(msg)
+
+    return list(map(dtype, lst))
+
+
+def interleave_iterables(iterable1: Iterable, iterable2: Iterable) -> List:
+    """Interleave two iterables element by element.
+
+    Parameters
+    ----------
+    iterable1
+        The first iterable to interleave.
+    iterable2
+        The second iterable to interleave.
+
+    Returns
+    -------
+    list
+        A new list with elements from `iterable1` and `iterable2` interleaved.
+
+    Raises
+    ------
+    TypeError
+        If either of the inputs is not an iterable of types: list, tuple,
+        string, or range.
+    ValueError
+        If the lengths of the two iterables do not match.
+
+    Examples
+    --------
+    >>> interleave_iterables([1, 2, 3], [4, 5, 6])
+    [1, 4, 2, 5, 3, 6]
+
+    >>> interleave_iterables((1, 2, 3), ('a', 'b', 'c'))
+    [1, 'a', 2, 'b', 3, 'c']
+
+    >>> interleave_iterables('ABC', '123')
+    ['A', '1', 'B', '2', 'C', '3']
+
+    >>> interleave_iterables(range(3), range(10, 13))
+    [0, 10, 1, 11, 2, 12]
+    """
+    if not isinstance(iterable1, (list, tuple, str, range)) or not isinstance(
+        iterable2,
+        (list, tuple, str, range),
+    ):
+        msg = (
+            "Both inputs must be iterable types such as list, tuple,"
+            "string, or range."
+        )
+        raise TypeError(msg)
+
+    if len(iterable1) != len(iterable2):
+        msg = "Both iterables must have the same length."
+        raise ValueError(msg)
+
+    result = [None] * (len(iterable1) + len(iterable2))
+    result[::2] = iterable1
+    result[1::2] = iterable2
+
+    return result
+
+
+def pairwise_iterable(iterable: Iterable) -> zip:
+    """Return pairs of adjacent values from the input iterable.
+
+    Parameters
+    ----------
+    iterable
+        An iterable object (e.g., list, tuple, string) from which pairs of
+        adjacent values will be generated.
+
+    Returns
+    -------
+    zip
+        An iterator of tuples, each containing a pair of adjacent values
+        from the input iterable.
+
+    Raises
+    ------
+    TypeError
+        If the input is not an iterable.
+
+    Examples
+    --------
+    >>> list(pairwise_iterable([1, 2, 3, 4]))
+    [(1, 2), (2, 3), (3, 4)]
+
+    >>> list(pairwise_iterable('abcde'))
+    [('a', 'b'), ('b', 'c'), ('c', 'd'), ('d', 'e')]
+
+    >>> list(pairwise_iterable((10, 20, 30)))
+    [(10, 20), (20, 30)]
+    """
+    if not hasattr(iterable, "__iter__"):
+        msg = "Input must be an iterable."
+        raise TypeError(msg)
+
+    a, b = tee(iterable)
+    next(b, None)
+    return zip(a, b)
+
+
+def merge_multi_dfs(
+    df_list: list,
+    on: Union[str, list],
+    how: str,
+    fillna_val: Union[None, object] = None,
+) -> pd.DataFrame:
+    """Perform consecutive merges on a list of pandas DataFrames.
+
+    Parameters
+    ----------
+    df_list
+        A list of DataFrames to be merged.
+    on
+        Column name(s) to merge on.
+    how
+        Type of merge to be performed. Must be one of 'left', 'right', 'outer',
+        'inner'.
+    fillna_val
+        Value to replace missing values with. Default is None.
+
+    Returns
+    -------
+    pd.DataFrame
+        The resulting DataFrame after merging and optional filling of missing
+        values.
+
+    Raises
+    ------
+    TypeError
+        If `df_list` is not a list of pandas DataFrames, or `on` is not a string
+        or list of strings, or `how` is not a string.
+    ValueError
+        If the `how` argument is not one of 'left', 'right', 'outer', or 'inner'.
+
+    Examples
+    --------
+    >>> import pandas as pd
+    >>> df1 = pd.DataFrame({'key': ['A', 'B', 'C'], 'value1': [1, 2, 3]})
+    >>> df2 = pd.DataFrame({'key': ['A', 'B'], 'value2': [4, 5]})
+    >>> df3 = pd.DataFrame({'key': ['A'], 'value3': [6]})
+    >>> merge_multi_dfs([df1, df2, df3], on='key', how='inner')
+      key  value1  value2  value3
+    0   A       1       4       6
+
+    >>> df1 = pd.DataFrame({'key': ['A', 'B', 'C'], 'value1': [1, 2, 3]})
+    >>> df2 = pd.DataFrame({'key': ['A', 'B'], 'value2': [4, 5]})
+    >>> merge_multi_dfs([df1, df2], on='key', how='outer',  fillna_val=0)
+      key  value1  value2
+    0   A        1        4
+    1   B        2        5
+    2   C        3        0
+    """
+    if not isinstance(df_list, list) or not all(
+        isinstance(df, pd.DataFrame) for df in df_list
+    ):
+        msg = "`df_list` must be a list of pandas DataFrames."
+        raise TypeError(msg)
+
+    if not isinstance(on, (str, list)):
+        msg = "`on` must be a string or a list of strings."
+        raise TypeError(msg)
+
+    if not isinstance(how, str):
+        msg = "`how` must be a string."
+        raise TypeError(msg)
+
+    valid_how_options = ["left", "right", "outer", "inner"]
+    if how not in valid_how_options:
+        msg = f"`how` Must be one of {valid_how_options}."
+        raise ValueError(msg)
+
+    merged_df = reduce(
+        lambda left, right: left.merge(right, on=on, how=how),
+        df_list,
+    )
+
+    if fillna_val is not None:
+        merged_df = merged_df.fillna(fillna_val)
+
+    return merged_df